Orginize the files new. Suppress the braces.
