+++ src/indent.cpp
@@ -20,6 +20,7 @@
 #include "align.h"
 #include "space.h"
 #include "parse_frame.h"
+#include "helper_for_print.h"
 
 
 /**
@@ -462,13 +463,11 @@ static void indent_pse_pop(parse_frame_t &frm, chunk_t *pc)
    {
       /* fatal error */
       fprintf(stderr, "the stack index is already zero\n");
-      {
-         char *outputMessage;
-         outputMessage = make_message("at line=%zu, type is %s\n",
-                                      pc->orig_line, get_token_name(pc->type));
-         fprintf(stderr, "%s", outputMessage);
-         free(outputMessage);
-      }
+      char *outputMessage;
+      outputMessage = make_message("at line=%zu, type is %s\n",
+                                   pc->orig_line, get_token_name(pc->type));
+      fprintf(stderr, "%s", outputMessage);
+      free(outputMessage);
       exit(EXIT_FAILURE);
    }
 } // indent_pse_pop
