+++ src/logger.cpp
@@ -444,43 +444,3 @@ void log_func_stack(log_sev_t sev, const char *prefix, const char *suffix, size_
       LOG_FMT(sev, "%s", suffix);
    }
 }
-
-
-char *make_message(const char *fmt, ...)
-{
-   size_t  n;
-   size_t  size = 100;  /* Guess we need no more than 100 bytes */
-   char    *p;
-   char    *np;
-   va_list ap;
-
-   if ((p = (char *)malloc(size)) == NULL)
-   {
-      return(NULL);
-   }
-
-   while (1)
-   {
-      /* Try to print in the allocated space */
-
-      va_start(ap, fmt);
-      n = static_cast<size_t>(vsnprintf(p, size, fmt, ap));
-      va_end(ap);
-
-      /* If that worked, return the string */
-      if (n < size)
-      {
-         return(p);
-      }
-
-      /* Else try again with more space */
-      size = n + 1;         /* Precisely what is needed */
-
-      if ((np = (char *)realloc(p, size)) == NULL)
-      {
-         free(p);
-         return(NULL);
-      }
-      p = np;
-   }
-} // make_message
